[Enh]: MpRelatableEntity - ppProject Extension
Need a trait based on MATPropertyOwner that is in a very low level project - maybe PharoEnhancements. This should be the foundation for Dynabook objects. Then the ProjectProperty trait can extend that trait. Right now, there doesn't seem to be a clean way to share this behavior with classes defined in projects that TPP depends on without creating a circular dependency.